style: group datasource change to inner class
