--- src/main/java/com/baomidou/dynamic/datasource/DynamicGroupDataSource.java
@@ -1,61 +1,0 @@
-/**
- * Copyright © 2018 organization baomidou
- * <pre>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * <pre/>
- */
-package com.baomidou.dynamic.datasource;
-
-import com.baomidou.dynamic.datasource.strategy.DynamicDataSourceStrategy;
-import lombok.Data;
-
-import javax.sql.DataSource;
-import java.util.LinkedList;
-import java.util.List;
-
-/**
- * 分组数据源
- *
- * @author TaoYu Kanyuxia
- * @since 2.0.0
- */
-@Data
-public class DynamicGroupDataSource {
-
-    private String groupName;
-
-    private DynamicDataSourceStrategy dynamicDataSourceStrategy;
-
-    private List<DataSource> dataSources = new LinkedList<>();
-
-    public DynamicGroupDataSource(String groupName, DynamicDataSourceStrategy dynamicDataSourceStrategy) {
-        this.groupName = groupName;
-        this.dynamicDataSourceStrategy = dynamicDataSourceStrategy;
-    }
-
-    public void addDatasource(DataSource dataSource) {
-        dataSources.add(dataSource);
-    }
-
-    public void removeDatasource(DataSource dataSource) {
-        dataSources.remove(dataSource);
-    }
-
-    public DataSource determineDataSource() {
-        return dynamicDataSourceStrategy.determineDataSource(dataSources);
-    }
-
-    public int size() {
-        return dataSources.size();
-    }
-}
